# opencode开发协作/templates/backend-springboot/src/main/resources/static/spec/openapi.yaml
openapi: 3.0.3
info:
  title: Enterprise AMis Platform Demo API
  version: 1.0.0

tags:
  - name: User
    description: User management
    x-menu:
      title: 用户管理
      icon: users
      path: /admin/users
      schemaRef: /schemas/user/list.json
      order: 10
    x-actions:
      - key: user:export
        title: 导出
      - key: user:resetPwd
        title: 重置密码

  - name: Approval
    description: Approval workflow
    x-menu:
      title: 审批中心
      icon: flow
      path: /flow/approvals
      schemaRef: /schemas/approval/list.json
      order: 20

  - name: Role
    description: Role & permissions
    x-menu:
      title: 角色权限
      icon: lock
      path: /sys/roles
      schemaRef: /schemas/rbac/role-list.json
      order: 90

  - name: Menu
    description: Menu configuration
    x-menu:
      title: 菜单管理
      icon: menu
      path: /sys/menus
      schemaRef: /schemas/menu/list.json
      order: 95

paths:
  /auth/me:
    get:
      tags: [User]
      summary: Current user & permissions
      x-permission: "auth:me"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiResponse_Me" }

  /dicts/{code}:
    get:
      tags: [User]
      summary: Dict items
      x-permission: "dict:query"
      parameters:
        - in: path
          name: code
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiResponse_DictItemArray" }

  /users:
    get:
      tags: [User]
      summary: Query users (paged)
      x-permission: "user:query"
      parameters:
        - in: query
          name: pageNo
          schema: { type: integer, default: 1 }
        - in: query
          name: pageSize
          schema: { type: integer, default: 20 }
        - in: query
          name: keyword
          schema: { type: string }
        - in: query
          name: menuId
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiResponse_PageResult_User" }

    post:
      tags: [User]
      summary: Create user
      x-permission: "user:create"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UserCreateRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiResponse_User" }

  /users/{id}:
    get:
      tags: [User]
      summary: User detail
      x-permission: "user:detail"
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiResponse_User" }

    put:
      tags: [User]
      summary: Update user
      x-permission: "user:update"
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UserUpdateRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiResponse_User" }

    delete:
      tags: [User]
      summary: Delete user
      x-permission: "user:delete"
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiResponse_Void" }

  /users/{id}/roles:
    get:
      tags: [User]
      summary: Get user roles
      x-permission: "user:role:read"
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiResponse_StringArray" }

    put:
      tags: [User]
      summary: Update user roles
      x-permission: "user:role:write"
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UserRolesUpdateRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiResponse_Void" }

  /roles:
    get:
      tags: [Role]
      summary: Query roles (paged)
      x-permission: "role:query"
      parameters:
        - in: query
          name: pageNo
          schema: { type: integer, default: 1 }
        - in: query
          name: pageSize
          schema: { type: integer, default: 20 }
        - in: query
          name: keyword
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiResponse_PageResult_Role" }

    post:
      tags: [Role]
      summary: Create role
      x-permission: "role:create"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RoleCreateRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiResponse_Role" }

  /roles/{id}/permissions:
    get:
      tags: [Role]
      summary: Get role permissions
      x-permission: "role:perm:read"
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiResponse_StringArray" }

    put:
      tags: [Role]
      summary: Update role permissions
      x-permission: "role:perm:write"
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RolePermissionsUpdateRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiResponse_Void" }

  /menus/tree:
    get:
      tags: [Menu]
      summary: Get menu tree for current user
      x-permission: "menu:tree"
      parameters:
        - in: query
          name: scene
          schema: { type: string, enum: [SIDEBAR, ADMIN] }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiResponse_MenuNodeArray" }

  /menus/{id}/actions:
    get:
      tags: [Menu]
      summary: Get enabled actions for a menu
      x-permission: "menu:action:read"
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiResponse_MenuActionArray" }
    put:
      tags: [Menu]
      summary: Update enabled actions for a menu
      x-permission: "menu:action:write"
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MenuActionsUpdateRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiResponse_Void" }

  /menus/{id}/field-policies:
    get:
      tags: [Menu]
      summary: Get field policies for a menu
      x-permission: "menu:field:read"
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiResponse_FieldPolicyArray" }
    put:
      tags: [Menu]
      summary: Update field policies for a menu
      x-permission: "menu:field:write"
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/FieldPoliciesUpdateRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiResponse_Void" }

  /menus/{id}/action-policies:
    get:
      tags: [Menu]
      summary: Get action policies for a menu
      x-permission: "menu:actionPolicy:read"
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiResponse_ActionPolicyArray" }
    put:
      tags: [Menu]
      summary: Update action policies for a menu
      x-permission: "menu:actionPolicy:write"
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ActionPoliciesUpdateRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiResponse_Void" }

components:
  schemas:
    ApiResponseBase:
      type: object
      required: [code, message, traceId]
      properties:
        code: { type: integer }
        message: { type: string }
        traceId: { type: string }
        data: {}

    ApiResponse_Void:
      allOf:
        - $ref: "#/components/schemas/ApiResponseBase"
        - type: object
          properties:
            data: { nullable: true }

    ApiResponse_StringArray:
      allOf:
        - $ref: "#/components/schemas/ApiResponseBase"
        - type: object
          properties:
            data:
              type: array
              items: { type: string }

    PageResultBase:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items: {}
        total: { type: integer }

    DictItem:
      type: object
      required: [label, value]
      properties:
        label: { type: string }
        value: { type: string }

    DictItemArray:
      type: array
      items: { $ref: "#/components/schemas/DictItem" }

    ApiResponse_DictItemArray:
      allOf:
        - $ref: "#/components/schemas/ApiResponseBase"
        - type: object
          properties:
            data: { $ref: "#/components/schemas/DictItemArray" }

    Me:
      type: object
      required: [id, name, permissions, defaultScope]
      properties:
        id: { type: string }
        name: { type: string }
        permissions:
          type: array
          items: { type: string }
        defaultScope:
          type: object
          properties:
            scopeType: { type: string }

    ApiResponse_Me:
      allOf:
        - $ref: "#/components/schemas/ApiResponseBase"
        - type: object
          properties:
            data: { $ref: "#/components/schemas/Me" }

    UserStatus:
      type: string
      enum: [ENABLED, DISABLED]
      x-dict: { type: remote, code: sys_user_status, api: /api/dicts/sys_user_status }

    User:
      type: object
      required: [id, username, realName, status, orgId, createdAt]
      properties:
        id: { type: string }
        username: { type: string }
        realName: { type: string }
        orgId: { type: string }
        status: { $ref: "#/components/schemas/UserStatus" }
        phone: { type: string, nullable: true, x-mask: { type: phone } }
        createdAt: { type: string, format: date-time }

    UserCreateRequest:
      type: object
      required: [username, realName, orgId, status]
      properties:
        username: { type: string }
        realName: { type: string }
        orgId: { type: string }
        status: { $ref: "#/components/schemas/UserStatus" }
        phone: { type: string, nullable: true }

    UserUpdateRequest:
      type: object
      properties:
        realName: { type: string }
        orgId: { type: string }
        status: { $ref: "#/components/schemas/UserStatus" }
        phone: { type: string, nullable: true }

    UserRolesUpdateRequest:
      type: object
      required: [roleIds]
      properties:
        roleIds:
          type: array
          items: { type: string }

    PageResult_User:
      allOf:
        - $ref: "#/components/schemas/PageResultBase"
        - type: object
          properties:
            items:
              type: array
              items: { $ref: "#/components/schemas/User" }

    ApiResponse_User:
      allOf:
        - $ref: "#/components/schemas/ApiResponseBase"
        - type: object
          properties:
            data: { $ref: "#/components/schemas/User" }

    ApiResponse_PageResult_User:
      allOf:
        - $ref: "#/components/schemas/ApiResponseBase"
        - type: object
          properties:
            data: { $ref: "#/components/schemas/PageResult_User" }

    Role:
      type: object
      required: [id, name]
      properties:
        id: { type: string }
        name: { type: string }
        code: { type: string, nullable: true }
        remark: { type: string, nullable: true }

    RoleCreateRequest:
      type: object
      required: [name]
      properties:
        name: { type: string }
        code: { type: string, nullable: true }
        remark: { type: string, nullable: true }

    RolePermissionsUpdateRequest:
      type: object
      required: [permissions]
      properties:
        permissions:
          type: array
          items: { type: string }

    PageResult_Role:
      allOf:
        - $ref: "#/components/schemas/PageResultBase"
        - type: object
          properties:
            items:
              type: array
              items: { $ref: "#/components/schemas/Role" }

    ApiResponse_Role:
      allOf:
        - $ref: "#/components/schemas/ApiResponseBase"
        - type: object
          properties:
            data: { $ref: "#/components/schemas/Role" }

    ApiResponse_PageResult_Role:
      allOf:
        - $ref: "#/components/schemas/ApiResponseBase"
        - type: object
          properties:
            data: { $ref: "#/components/schemas/PageResult_Role" }

    MenuNode:
      type: object
      required: [id, title, type]
      properties:
        id: { type: string }
        title: { type: string }
        type: { type: string }
        path: { type: string, nullable: true }
        schemaRef: { type: string, nullable: true }
        orderNo: { type: integer, nullable: true }
        children:
          type: array
          items: { $ref: "#/components/schemas/MenuNode" }

    MenuNodeArray:
      type: array
      items: { $ref: "#/components/schemas/MenuNode" }

    ApiResponse_MenuNodeArray:
      allOf:
        - $ref: "#/components/schemas/ApiResponseBase"
        - type: object
          properties:
            data: { $ref: "#/components/schemas/MenuNodeArray" }

    MenuAction:
      type: object
      required: [key, title, enabled]
      properties:
        key: { type: string }
        title: { type: string }
        enabled: { type: boolean }

    MenuActionArray:
      type: array
      items: { $ref: "#/components/schemas/MenuAction" }

    ApiResponse_MenuActionArray:
      allOf:
        - $ref: "#/components/schemas/ApiResponseBase"
        - type: object
          properties:
            data: { $ref: "#/components/schemas/MenuActionArray" }

    MenuActionsUpdateRequest:
      type: object
      required: [actions]
      properties:
        actions:
          type: array
          items:
            type: object
            required: [key, enabled]
            properties:
              key: { type: string }
              enabled: { type: boolean }

    FieldPolicyTarget:
      type: string
      enum: [LIST, DETAIL, FORM]
    FieldPolicyMode:
      type: string
      enum: [VISIBLE, HIDDEN, READONLY, MASKED]

    FieldPolicy:
      type: object
      required: [resource, field, target, mode]
      properties:
        resource: { type: string }
        field: { type: string }
        target: { $ref: "#/components/schemas/FieldPolicyTarget" }
        mode: { $ref: "#/components/schemas/FieldPolicyMode" }
        when: { type: string, nullable: true }
        priority: { type: integer, nullable: true }
        maskType: { type: string, nullable: true }
        unmaskPermission: { type: string, nullable: true }

    FieldPolicyArray:
      type: array
      items: { $ref: "#/components/schemas/FieldPolicy" }

    ApiResponse_FieldPolicyArray:
      allOf:
        - $ref: "#/components/schemas/ApiResponseBase"
        - type: object
          properties:
            data: { $ref: "#/components/schemas/FieldPolicyArray" }

    FieldPoliciesUpdateRequest:
      type: object
      required: [policies]
      properties:
        policies:
          type: array
          items: { $ref: "#/components/schemas/FieldPolicy" }

    ActionPolicyTarget:
      type: string
      enum: [TOOLBAR, ROW, DETAIL]
    ActionPolicyMode:
      type: string
      enum: [ALLOW, HIDE, DISABLE]

    ActionPolicy:
      type: object
      required: [resource, actionKey, target, mode]
      properties:
        resource: { type: string }
        actionKey: { type: string }
        target: { $ref: "#/components/schemas/ActionPolicyTarget" }
        mode: { $ref: "#/components/schemas/ActionPolicyMode" }
        when: { type: string, nullable: true }
        priority: { type: integer, nullable: true }
        disabledReason: { type: string, nullable: true }
        confirmText: { type: string, nullable: true }
        requireReason: { type: boolean, nullable: true }

    ActionPolicyArray:
      type: array
      items: { $ref: "#/components/schemas/ActionPolicy" }

    ApiResponse_ActionPolicyArray:
      allOf:
        - $ref: "#/components/schemas/ApiResponseBase"
        - type: object
          properties:
            data: { $ref: "#/components/schemas/ActionPolicyArray" }

    ActionPoliciesUpdateRequest:
      type: object
      required: [policies]
      properties:
        policies:
          type: array
          items: { $ref: "#/components/schemas/ActionPolicy" }